requirements.txt
tensorflow-cpu
nltk
numpy